+++ src/components/cashconnect/CCSessionProposalDialog.vue
